Put back world-offset, (but left commented lines to let J tweak)

+++ src/afk/ge/tokyo/ems/components/HUDTag.java
@@ -6,8 +6,8 @@ package afk.ge.tokyo.ems.components;
  */
 public class HUDTag
 {
-    public int x = 0;
-    public int y = 0;
+    public float x = 0;
+    public float y = 0;
     
     public boolean centerX = false;
     public boolean centerY = false;
@@ -16,7 +16,7 @@ public class HUDTag
     {
     }
 
-    public HUDTag(int x, int y, boolean centerX, boolean centerY)
+    public HUDTag(float x, float y, boolean centerX, boolean centerY)
     {
         this.x = x;
         this.y = y;
